+++ packages/core/src/components/graphs/bar-grouped.ts
@@ -1,5 +1,6 @@
 import { ScaleBand, scaleBand, select } from 'd3'
-import { generateSVGPathString, getProperty, removeArrayDuplicates } from '@/tools'
+import { uniq } from 'lodash-es'
+import { generateSVGPathString, getProperty } from '@/tools'
 import { Bar } from './bar'
 import {
 	CartesianOrientations,
@@ -42,7 +43,7 @@ export class GroupedBar extends Bar {
 		// Grab container SVG
 		const svg = this.getComponentContainer({ withinChartClip: true })
 
-		const allDataLabels = removeArrayDuplicates(
+		const allDataLabels = uniq(
 			displayData.map((datum: any) => {
 				const domainIdentifier = this.services.cartesianScales.getDomainIdentifier(datum)