+++ packages/core/src/demo/utils/stackblitz/svelte.ts
@@ -1,14 +1,12 @@
 import type { Project, ProjectTemplate } from '@stackblitz/sdk'
 import type { Demo } from '@/demo'
-
+import { objectToString } from './object-to-string'
 import { version } from '../package-versions'
 
 export function buildSvelteExample(demo: Demo): Project {
 
   const devDependencies: Record<string, string> = {
-    '@carbon/charts': version.carbonCharts,
     '@carbon/charts-svelte': version.carbonCharts,
-    '@carbon/styles': version.carbonStyles,
     '@sveltejs/vite-plugin-svelte': version.svelteVite,
     '@tsconfig/svelte': version.svelteTsConfig,
     d3: version.d3,
@@ -37,10 +35,9 @@ export function buildSvelteExample(demo: Demo): Project {
   const appSvelte =
 `<script lang="ts">
 import { ${chartComponent} } from '@carbon/charts-svelte'
-import '@carbon/styles/css/styles.css'
 import '@carbon/charts-svelte/styles.css'
-import options from './options'
-import data from './data.json'
+import options from './options.js'
+import data from './data.js'
 </script>
 
 <${chartComponent} {data} {options} style="padding:2rem;" />
@@ -137,11 +134,7 @@ export default {
 import { svelte } from '@sveltejs/vite-plugin-svelte'
 
 export default defineConfig({
-	plugins: [svelte()],
-	optimizeDeps: {
-		include: ['@carbon/charts', 'carbon-components'],
-		exclude: ['@carbon/telemetry']
-	}
+	plugins: [svelte()]
 })
 `
 
@@ -150,9 +143,9 @@ export default defineConfig({
     title: 'Carbon Charts Svelte Example',
     files: {
       'src/App.svelte': appSvelte,
-      'src/data.json': JSON.stringify(demo.data, null, 2),
+      'src/data.js': objectToString(demo.data),
       'src/main.ts': mainTs,
-      'src/options.json': JSON.stringify(demo.options, null, 2),
+      'src/options.js': objectToString(demo.options),
       'src/vite-env.d.ts': viteEnvDts,
       '.stackblitzrc': stackBlitzRc,
       'index.html': indexHtml,