+++ packages/core/src/tests/index.ts
@@ -1,1 +1,2 @@
 export { data, options, TestEnvironment } from './test-environment'
+export const prefix = 'bx'