+++ packages/svelte/src/lib/ChoroplethChart.svelte
@@ -6,10 +6,10 @@
 	} from '@carbon/charts'
 	import BaseChart from './BaseChart.svelte'
 
-	export let chart: ChoroplethChartCore
 	export let options: ChoroplethChartOptions
 	export let data: ChartTabularData
-	export let ref: HTMLDivElement
+	export let chart: ChoroplethChartCore | null  = null
+	export let ref: HTMLDivElement | null  = null
 </script>
 
 <BaseChart