+++ packages/svelte/src/lib/CirclePackChart.svelte
@@ -6,10 +6,10 @@
 	} from '@carbon/charts'
 	import BaseChart from './BaseChart.svelte'
 
-	export let chart: CirclePackChartCore
 	export let options: CirclePackChartOptions
 	export let data: ChartTabularData
-	export let ref: HTMLDivElement
+	export let chart: CirclePackChartCore | null  = null
+	export let ref: HTMLDivElement | null  = null
 </script>
 
 <BaseChart